Refactor the index into its own module
